Completed: lifted state to show location on map, separated location input, updated state. #5
