+++ client/components/location-input.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export default function LocationInput(props) {
+
+  return (
+      <input
+        required
+        type="text"
+        name="location"
+        id="location"
+        onChange={props.handleChange}
+        value={props.location}
+        placeholder="Search Location"
+        className="form-control"
+      />
+  );
+}
